Extract Highlight related methods into a service class
